+++ chapter_05/exercise_5_5/strncpy.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 
-void strcpy_ptr(char *s, char *t, unsigned int n);
+void strcpy_ptr(char *s, char *t, size_t n);
 
 int main(void)
 {
   char s[100] = "This is the first string";
   char *t = "Test is the second string, that is cool";
-  unsigned int nr_chars = 26;
+  size_t nr_chars = 26;
 
   strcpy_ptr(s, t, nr_chars);
 
@@ -15,7 +15,7 @@ int main(void)
   return 0;
 }
 
-void strcpy_ptr(char *s, char *t, unsigned int n)
+void strcpy_ptr(char *s, char *t, size_t n)
 {
   while ((*s++ = *t++) && --n)
     ;
